move embedding backends
