+++ haystack/preview/components/embedders/backends/sentence_transformers_backend.py
