+++ test/preview/components/embedders/test_sentence_transformers_embedding_backend.py
@@ -1,12 +1,12 @@
 from unittest.mock import patch
 import pytest
-from haystack.preview.embedding_backends.sentence_transformers_backend import (
+from haystack.preview.components.embedders.backends.sentence_transformers_backend import (
     _SentenceTransformersEmbeddingBackendFactory,
 )
 
 
 @pytest.mark.unit
-@patch("haystack.preview.embedding_backends.sentence_transformers_backend.SentenceTransformer")
+@patch("haystack.preview.components.embedders.backends.sentence_transformers_backend.SentenceTransformer")
 def test_factory_behavior(mock_sentence_transformer):
     embedding_backend = _SentenceTransformersEmbeddingBackendFactory.get_embedding_backend(
         model_name_or_path="my_model", device="cpu"
@@ -21,7 +21,7 @@ def test_factory_behavior(mock_sentence_transformer):
 
 
 @pytest.mark.unit
-@patch("haystack.preview.embedding_backends.sentence_transformers_backend.SentenceTransformer")
+@patch("haystack.preview.components.embedders.backends.sentence_transformers_backend.SentenceTransformer")
 def test_model_initialization(mock_sentence_transformer):
     _SentenceTransformersEmbeddingBackendFactory.get_embedding_backend(
         model_name_or_path="model", device="cpu", use_auth_token="my_token"
@@ -32,7 +32,7 @@ def test_model_initialization(mock_sentence_transformer):
 
 
 @pytest.mark.unit
-@patch("haystack.preview.embedding_backends.sentence_transformers_backend.SentenceTransformer")
+@patch("haystack.preview.components.embedders.backends.sentence_transformers_backend.SentenceTransformer")
 def test_embedding_function_with_kwargs(mock_sentence_transformer):
     embedding_backend = _SentenceTransformersEmbeddingBackendFactory.get_embedding_backend(model_name_or_path="model")
